// src/types/jumpstart.ts
export type JumpstartThemes =
  | "Basri"
  | "Unicorns"
  | "Teferi"
  | "Mill"
  | "Liliana"
  | "Phyrexian"
  | "Chandra"
  | "Seismic"
  | "Garruk"
  | "Walls"
  | "Rainbow"
  | "Angels"
  | "Dogs"
  | "Enchanted"
  | "Pirates"
  | "Spirits"
  | "Under the Sea"
  | "Discarding"
  | "Rogues"
  | "Witchcraft"
  | "Dragons"
  | "Lightning"
  | "Minotaurs"
  | "Cats"
  | "Elves"
  | "Lands"
  | "Doctor"
  | "Feathered Friends"
  | "Heavily Armored"
  | "Legion"
  | "Above the Clouds"
  | "Archaeology"
  | "Well-Read"
  | "Wizards"
  | "Minions"
  | "Reanimated"
  | "Spooky"
  | "Vampires"
  | "Devilish"
  | "Goblins"
  | "Smashing"
  | "Spellcasting"
  | "Dinosaurs"
  | "Plus One"
  | "Predatory"
  | "Tree-Hugging";